continued work on restructuring template
