+++ app/assets/styles/layout.css
@@ -1,0 +1,3 @@
+body {
+  font-family: cursive;
+}
