+++ app/store/store.js
@@ -1,23 +1,23 @@
-// import { createStore, applyMiddleware } from 'redux';
-// import thunk from 'redux-thunk';
-//
-// import rootReducer from './reducers';
-//
-// const createStoreWithMiddleware = applyMiddleware(
-//   thunk
-// )(createStore);
-//
-// export default function configureStore(initialState) {
-//   const store = createStoreWithMiddleware(rootReducer, initialState);
-//   if (module.hot) {
-//     // Enable Webpack hot module replacement for reducers
-//     module.hot.accept('./reducers', () => {
-//       var nextRootReducer = require('./reducers');
-//       store.replaceReducer(nextRootReducer);
-//     });
-//   }
-//   return store;
-// }
+import { createStore, applyMiddleware } from 'redux';
+import thunk from 'redux-thunk';
+
+import rootReducer from './reducers';
+
+const createStoreWithMiddleware = applyMiddleware(
+  thunk
+)(createStore);
+
+export default function configureStore(initialState) {
+  const store = createStoreWithMiddleware(rootReducer, initialState);
+  if (module.hot) {
+    // Enable Webpack hot module replacement for reducers
+    module.hot.accept('./reducers', () => {
+      var nextRootReducer = require('./reducers');
+      store.replaceReducer(nextRootReducer);
+    });
+  }
+  return store;
+}
