+++ app/views/pages/HomePage/index.js
@@ -1,16 +1,16 @@
 import React from 'react';
 
-// import styles from './styles.css';
+import styles from './styles.css';
 
 export const homePage = React.createClass({
 
   render: function () {
-    // return (
-    //   <h1 className={styles.h1}>Hello, world!</h1>
-    // );
     return (
-      <h1>Hello, world!</h1>
+      <h1 className={styles.header}>Hello, world!</h1>
     );
+    // return (
+    //   <h1>Hello, world!</h1>
+    // );
   },
 
 });
