--- client/common/layout.css
@@ -1,3 +1,0 @@
-.box {
-    padding: 10px;
-}
