--- client/components/App.css
@@ -1,8 +1,0 @@
-.increment {
-    composes: box from "../common/layout.css";
-    color: white;
-    cursor: pointer;
-    border: 1px solid green;
-    border-radius: 3px;
-    background-color: red;
-}
