+++ client/entry.js
@@ -1,3 +1,5 @@
+/* eslint-env browser */
+import 'babel-core/register';
 import 'babel-polyfill';
 
 import React from 'react';
