--- client/reducers/index.js
@@ -1,17 +1,0 @@
-var initial = {
-  num: 5,
-};
-var handlers = {
-  INC(s) {
-    s.num += 1;
-    return s;
-  }
-};
-function reducer(state = initial, action) {
-  if (handlers[action.type]) {
-    return handlers[action.type](state);
-  }
-  return state;
-}
-
-module.exports = reducer;
