--- index.html
@@ -1,13 +1,0 @@
-<html>
-  <head>
-    <title>Sample App</title>
-  </head>
-  <body>
-    <div id="root" data-js="app"><!-- CONTENT --></div>
-  </body>
-<script type="text/javascript">
-/* eslint-env browser */
-window.initialStoreData = "-- STORES --";
-</script>
-  <script src="/scripts/dev_bundle.js"></script>
-</html>
