--- server/app.js
@@ -1,9 +1,0 @@
-var express = require('express');
-
-var app = express.Router();
-
-app.get('/whoami', (req, res) => {
-  res.send("You are a winner");
-});
-
-module.exports = app;
